+++ display.rb
@@ -1,0 +1,10 @@
+def display_menu
+    puts 'Library App Menu:'
+    puts '1. List all books'
+    puts '2. List all people'
+    puts '3. Create a person'
+    puts '4. Create a book'
+    puts '5. Create a rental'
+    puts '6. List all rentals for a given person id'
+    puts '7. Quit the app'
+  end
