+++ main.rb
@@ -1,4 +1,5 @@
 require './app'
+require './display'
 class Main
   def initialize
     @app = App.new
@@ -34,16 +35,16 @@ class Main
     puts 'Thank you for using the Library App. Goodbye!'
   end
 
-  def display_menu
-    puts 'Library App Menu:'
-    puts '1. List all books'
-    puts '2. List all people'
-    puts '3. Create a person'
-    puts '4. Create a book'
-    puts '5. Create a rental'
-    puts '6. List all rentals for a given person id'
-    puts '7. Quit the app'
-  end
+  # def display_menu
+  #   puts 'Library App Menu:'
+  #   puts '1. List all books'
+  #   puts '2. List all people'
+  #   puts '3. Create a person'
+  #   puts '4. Create a book'
+  #   puts '5. Create a rental'
+  #   puts '6. List all rentals for a given person id'
+  #   puts '7. Quit the app'
+  # end
 end
 
 main = Main.new
